ai_matching_system/requirements.txt
# AI Matching System Requirements
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
typing-extensions>=4.9.0
tavily-python>=0.2.0
numpy>=1.24.0
pandas>=2.0.0
aiofiles>=23.0.0
pinecone>=2.0.0
tqdm>=4.65.0
supabase>=2.0.0